feat(gazelle): support multiple requirements files in manifest generation (#1301)
For certain workflows it is useful to calculate the integrity hash of
the manifest file based on a number of requirements files. The
requirements locking is usually done by executing a script on each
platform and having gazelle manifest generator be aware that more than
one requirements file may affect the outcome (e.g. the wheels that get
passed to modules map may come from multi_pip_parse rule) is generally
useful.

This change modifies the generation macro to concatenate the
requirements
files into one before passing it to the manifest generator.