build(gazelle): embed Python zip file (#1485)
The runtime dependencies of Gazelle Python extension makes it hard to
distribute Gazelle binaries: we have to preserve the runfiles structure
and distribute it with Gazelle binaries.

Instead, we can build a single Python zip file that comes a built-in
interpreter, and embed the zip file into the Go binary in compile time
and avoid the runtime dependency.

Fixes #1455

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>