feat(gazelle): use relative paths for resolved imports (#1554)
Modify the Gazelle plugin so that when it adds a `dep` because of a
`resolve` directive it makes it a relative import if possible.

The first commit adds a test for the existing behavior, where inside of
`//package2` the dependency `//package2:resolved_package` is added. The
second commit updates the test and the behavior so inside of
`//package2` we add `: resolved_package` instead.

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>