feat(gazelle): allow per-file py_test generation (#1563)
Previously the per-file target generation only worked for py_library
targets. This change makes it so that this feature works for py_test
targets as well.
The change is careful to not affect any existing tests, so I'm not sure
if it should count as a breaking change. New tests have been added to
check the new functionality.